Breakout servo reboot, remove velocity limit
Getting ready to add a command to reset the servo if it over torques during use, move the servo setup out it's own and call it from the main setup code. 

Remove the velocity limit so we can open and close the vice faster. If this causes problems we can just move to a faster velocity limit.