Include stepper support, make servo support optional.
defines STEPPER_SUPPORT and SERVO_SUPPORT can be in or commented out to put in those parts. As before, if servos are supported, the shield is required and serial communication must be done through an external USB -> Serial adapter as documented. Without that option, we should be able to just use the built in Arduino interface. 
Stepper motor support allows setup and operation via:
#,#M //stepper Motor. <step_pin>,<dir_pin>M. Defaults to pins 3 and 4
#,#V //motion profile for the stepper. <acceleration>, <velocity limit>V. Defaults to 240 and 480 steps per second.
#G   //goto. Move the stepper motor to the specified position.

See
https://www.airspayce.com/mikem/arduino/AccelStepper/index.html for details. Install library via 
"Sketch" / "Include Libraries" / "Manage Libraries" and then search for "AccelStep". Select and then "Install"