Bug: ETO vs EOT, Report error if servo doesn't connect
This wasn't using the EOT option at all because the stupid define was set to ETO. Dyslexia for the win!

Also added code to return an error if the servo doesn't respond. It would say that the velocity couldn't be set before, but that was rather misleading when the actual cause was that the servo wasn't responding at all.