BUG: Don't clear negative sign until all digits are processed.
Facepalm. For negative values, it's critical that we keep each digit going in as a negative until we stop getting digits. I was (stupidly) clearing the sign after the first digit so any following digits would be added to the (negative) total. e.g. -12 would end up as -8 because that is -10 + 2. Anyway. Works for negative values now.