I and P now replace n with pin or register value
Digital ONLY pins are now differentiated from analog pins. e.g. on Arduino Uno, 2 to 13 are digital only, 14 to 19 are analog (or digital) but we read them via analog. This allows I and P to be used to read values inside a program, since ? is a conditional branch. e.g. 2I will read pin 2 and set n to 0 or 1. 14I will read pin 14 aka A0 and set n to the analog value read on the pin.